Api handling add json exception check, fix model

+++ hse_mooc/app/src/main/java/ru/andrewkir/hse_mooc/network/responses/CoursesPreview/External.kt
@@ -6,6 +6,6 @@ import android.os.Parcelable
 
 @Parcelize
 data class External(
-    val averageScore: Int,
+    val averageScore: Double,
     val countReviews: Int
 ) : Parcelable
